fix(website): address cubic review findings on PR #275
- analytics: init posthog inside loadPosthog() to avoid capturing
  before init, and reset the cached promise on load failure
- Features: add aria-label/type to icon-only send button; extract
  duplicated Steno mic-icon SVG into StenoMicIcon
- Models: drop unused ScrollText import; derive "N models" copy from
  localModels.length instead of hardcoding "Five"
- extract duplicated RecordingWave into shared components/RecordingWave.jsx
